PE-0045: Answer accepted.

+++ Demo.ProjectEuler/Demo.ProjectEuler.Core/NumberGenerator.cs
@@ -3,19 +3,19 @@ namespace Demo.ProjectEuler.Core
 	public class NumberGenerator
 	{
 		/// <param name="n">One-based index</param>
-		public int GetTriangleNumber(int n)
+		public long GetTriangleNumber(long n)
 		{
 			return n * (n + 1) / 2;
 		}
 
 		/// <param name="n">One-based index</param>
-		public int GetPentagonalNumber(int n)
+		public long GetPentagonalNumber(long n)
 		{
 			return n * (3 * n - 1) / 2;
 		}
 
 		/// <param name="n">One-based index</param>
-		public int GetHexagonalNumber(int n)
+		public long GetHexagonalNumber(long n)
 		{
 			return n * (2 * n - 1);
 		}
